+++ sources/sections/objectMatrix.c
@@ -27,7 +27,7 @@ void preload_objectMatrix () {
 void load_objectMatrix() {
 	char *evalEquation;
 	char *evalSources;
-	int equationSize;
+	size_t equationSize;
 	H3dsMeshObj *mesh;
 	int i;
